feat(ui): Add Related Identifier Information to Credential Details (#839)
* fix(ui): prevent submit when input invalid name

* feat(ui): I create identifier detail module

* feat(ui): update identifier detail

* feat(ui): add related identifier to cred

* fix(ui): Fix duplicate connection

* feat(ui): hide related identifier when cannot find identifier

---------
