+++ src/ui/components/IdentifierDetailModule/IdentifierDetailModal.tsx
@@ -1,0 +1,46 @@
+import { IonModal } from "@ionic/react";
+import { useCallback, useMemo } from "react";
+import { BackEventPriorityType } from "../../globals/types";
+import { IdentifierDetailModule } from "./IdentifierDetailModule";
+import { IdentifierDetailModalProps } from "./IdentifierDetailModule.types";
+
+const IdentifierDetailModal = ({
+  isOpen,
+  setIsOpen,
+  onClose,
+  ...props
+}: IdentifierDetailModalProps) => {
+  const handleClose = useCallback(() => {
+    setIsOpen(false);
+  }, [setIsOpen]);
+
+  const handleBack = useCallback(() => {
+    handleClose();
+    onClose?.();
+  }, [handleClose, onClose]);
+
+  const hardwareBackButtonConfig = useMemo(
+    () => ({
+      prevent: false,
+      priority: BackEventPriorityType.Modal,
+    }),
+    []
+  );
+
+  return (
+    <IonModal
+      data-testid={`${props.pageId}-modal`}
+      isOpen={isOpen}
+      onIonModalDidDismiss={handleClose}
+    >
+      <IdentifierDetailModule
+        {...props}
+        onClose={handleBack}
+        hardwareBackButtonConfig={hardwareBackButtonConfig}
+        navAnimation={false}
+      />
+    </IonModal>
+  );
+};
+
+export { IdentifierDetailModal };
