+++ src/ui/components/IdentifierDetailModule/IdentifierDetailModule.scss
@@ -1,0 +1,165 @@
+.identifier-details-module {
+  .identifier-card-template {
+    margin-bottom: 0;
+  }
+
+  .card-details-content .card-block {
+    margin: 0.75rem 0 0;
+
+    &.flat-border-top {
+      margin-top: 0.25rem;
+    }
+  }
+
+  .creation-timestamp {
+    .card-details-info-block-data {
+      font-size: 0.875rem;
+      font-weight: 500;
+      color: var(--ion-color-secondary);
+      white-space: nowrap;
+
+      .card-details-info-block-key {
+        display: block;
+        font-size: 1rem;
+        color: var(--ion-color-primary);
+      }
+    }
+  }
+
+  .list-header {
+    margin: 1.5rem 0 0 !important;
+  }
+
+  .member:not(:last-child) {
+    margin-bottom: 0.75rem;
+  }
+
+  .view-more-members {
+    --background: rgba(var(--ion-color-dark-grey-rgb), 0.5);
+    --border-radius: 0.5rem;
+    color: var(--ion-color-primary);
+    width: 100%;
+    display: block;
+    margin-top: 1.25rem;
+    --box-shadow: none;
+    --padding-top: 0.625rem;
+    --padding-bottom: 0.625rem;
+    font-size: 1rem;
+    --background-activated: rgba(var(--ion-color-dark-grey-rgb), 0.5);
+
+    &::part(native) {
+      border: none;
+    }
+  }
+
+  .identifier-details-split-section {
+    margin: 0 !important;
+    display: flex;
+    justify-content: space-between;
+    align-items: stretch;
+
+    .card-block {
+      width: 49%;
+
+      .card-details-info-block-inner {
+        height: 100%;
+      }
+    }
+  }
+
+  .rotate-keys-button {
+    width: 100%;
+    --background: transparent;
+    --box-shadow: none;
+    margin: 0;
+    --background-activated: transparent;
+    min-height: auto;
+
+    p {
+      margin: 0 0.25rem 0 0;
+    }
+
+    p,
+    ion-icon {
+      color: var(--ion-color-action-blue);
+    }
+  }
+
+  .page-header {
+    margin-top: 0.625rem;
+  }
+
+  @media screen and (min-width: 250px) and (max-width: 370px) {
+    .identifier-details-split-section {
+      flex-direction: column;
+
+      .card-block {
+        width: auto;
+      }
+    }
+
+    .creation-timestamp {
+      .card-details-info-block-data {
+        font-size: 0.75rem;
+
+        .card-details-info-block-key {
+          font-size: 0.8rem;
+        }
+      }
+    }
+
+    .view-more-members {
+      font-size: 0.8rem;
+    }
+  }
+}
+
+.open-animation {
+  .card-details-content {
+    animation: expand 0.7s ease-in-out;
+    margin-top: 0 !important;
+
+    @keyframes expand {
+      from {
+        height: 0;
+        overflow-y: hidden;
+        opacity: 0;
+      }
+
+      to {
+        height: 100vh;
+        overflow-y: hidden;
+        opacity: 1;
+      }
+    }
+  }
+}
+
+.back-animation {
+  .card-details-content {
+    animation: collapse 0.2s ease-in-out;
+    animation-fill-mode: forwards;
+    margin-top: 0 !important;
+  }
+
+  @keyframes collapse {
+    from {
+      height: 100vh;
+      overflow-y: hidden;
+      opacity: 1;
+    }
+
+    to {
+      height: 0;
+      overflow-y: hidden;
+      opacity: 0;
+    }
+  }
+}
+
+.identifier-card-detail-spinner-container {
+  height: calc(100vh - 11.5rem);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
