+++ src/ui/components/IdentifierDetailModule/IdentifierDetailModule.types.ts
@@ -1,0 +1,16 @@
+import { HardwareBackButtonConfig } from "../PageHeader/PageHeader.types";
+
+interface IdentifierDetailModuleProps {
+    pageId: string;
+    identifierDetailId: string;
+    onClose?: (animation?: boolean) => void;
+    navAnimation: boolean;
+    hardwareBackButtonConfig?: HardwareBackButtonConfig;
+}
+
+interface IdentifierDetailModalProps extends Omit<IdentifierDetailModuleProps, "navAnimation"> {
+    isOpen: boolean;
+    setIsOpen: (value: boolean) => void;
+}
+
+export type { IdentifierDetailModuleProps, IdentifierDetailModalProps };
