+++ src/ui/components/IdentifierDetailModule/components/IdentifierAttributeDetailModal/Advanced.tsx
