+++ src/ui/components/IdentifierDetailModule/components/IdentifierAttributeDetailModal/IdentifierAttributeDetailModal.tsx
