+++ src/ui/components/IdentifierDetailModule/components/IdentifierAttributeDetailModal/IdentifierAttributeDetailModal.types.ts
