+++ src/ui/components/IdentifierDetailModule/components/IdentifierAttributeDetailModal/List.tsx
