+++ src/ui/components/IdentifierDetailModule/components/IdentifierAttributeDetailModal/RotationThreshold.tsx
