+++ src/ui/components/IdentifierDetailModule/components/IdentifierContent.types.ts
