+++ src/ui/components/IdentifierDetailModule/components/RotateKeyModal/RotateKeyModal.scss
