+++ src/ui/components/IdentifierDetailModule/components/RotateKeyModal/RotateKeyModal.types.ts
