+++ src/ui/components/IdentifierDetailModule/components/index.ts
