+++ src/ui/components/IdentifierDetailModule/index.ts
@@ -1,0 +1,2 @@
+export * from "./IdentifierDetailModule";
+export * from "./IdentifierDetailModal";
