+++ src/ui/pages/Credentials/Credentials.scss
@@ -111,7 +111,7 @@
     }
 
     .credential-cards {
-      .list-header {
+      .list-header,.credentials-tab-chips {
         animation: indentifierFavouriteCardHeaderResize 0.25s ease;
         animation-fill-mode: forwards;
