+++ src/ui/pages/NotificationDetails/components/CredentialRequest/ChooseCredential/ChooseCredential.test.tsx
@@ -254,6 +254,9 @@ describe("Credential request - choose request", () => {
           { ...credsFixAcdc[0], id: credRequestFix.credentials[0].acdc.d },
         ],
       },
+      identifiersCache: {
+        identifiers: [],
+      },
     };
 
     const storeMocked = {
